Bug and Glitch Fixes
Addressed bug where new terms would not be added to Glossary table if not enough rows were present. Also, added code to remove extra rows in Glossary table when more rows than terms are present in the table.

Changed all instances of frozenset to set due to usage of frozensets being deemed unnecessary.

Added check so that sym.renderer.groups does not throw an error due to not having any symbology set up for MapUnitPolys feature class(es) if the user enables this for the tool.