Fixes to AutofillAreaFillRGB and Symbol feature
Autofill AreaFillRGB and Symbol feature of GeMS Auto-Filler and Fixer tool now will obtain Symbology information from MapUnitOverlayPolys in the GeologicMap dataset and add them to DescriptionOfMapUnits table (excluding Qal and water, which are still handled and treated separately). An abnormal (and assumingly rare) glitch was addressed in terms of how the tool obtains Symbology information from MapUnitPolys. A fix was added that prevents arcpy from reading Symbology information that was remove from the Symbology of the feature class.