Major Update
Fixes for GeMS Auto-Filler and Fixer tool:

Autofill AreaFillRGB and Symbol will no longer cause the tool to fail if Annotation feature types are present in the GeMS geodatabase.

Autofill Label and Symbols fields of polygon feature classes will no longer fail if OverlayPolys feature class is present in the GeMS geodatabase. OverlyPolys will be ignored when this is enabled for the tool now.


Changes:

Slight optimizations to GeMS Auto-Filler and Fixer tool and GeMS Geometry Fixer tool.