Significant Changes and Improvements to Tools
Properly implemented automated explicit typo errors that will always execute when using the tools.

Implemented automated consistency text enforcer that will always execute when using the GeMS Auto-Filler and Fixer tool as well as the GeMS Geometry Fixer tool. For example, say in the isConcealed field for MapUnitLines uses both "n" and "N" to indicate the feature is not concealed. This newly implemented algorithm will make it so that, for example, all entries in the isConcealed field for MapUnitLines are lowercased so only "n" appears instead of sometimes being represented as "N". This is also contextually based for each feature class.

A new option for the  tool will be added for specifically and contextually dealing with the Label field in feature classes. For example, for OrientationPoints, it ensures that the Inclination and Label fields for every point are referring to the same number. If the Symbol field is specified as "hidden", the Label field for the related point will be set to None (i.e., disable as <Null> indicating no data). For another example, for MapUnitPolys, the MapUnit, Symbol, and Label fields will attempted to be matched with the corresponding information present in the DescriptionOfMapUnits table. Like all the other functions in GeMS Auto-Filler and Fixer tool, no changes will be made to anything within the geodatabase unless necessary and certain.

GeMS Geometry Fixer is planned to be expanded and fully reworked in the future so it has similar utility (but not function) to GeMS Auto-Filler and Fixer as of right now it only merges adjacent polygons of matching information.